moved Configuration class into 'api.config' package
